优化移除容器的功能

+++ manager.sh
@@ -1220,23 +1220,28 @@ remove_gnu_linux_container() {
 		echo "${YELLOW}It is detected that you do not currently have GNU/Linux container installed. 检测到您当前未安装容器${RESET}"
 	fi
 	echo "${YELLOW}按回车键确认移除,按Ctrl+C取消 Press enter to confirm.${RESET} "
-	read
-
-	chmod 777 -R ${DEBIAN_FOLDER}
-	rm -rfv "${DEBIAN_FOLDER}" ${PREFIX}/bin/debian ${PREFIX}/bin/startvnc ${PREFIX}/bin/stopvnc ${PREFIX}/bin/startxsdl ${PREFIX}/bin/debian-rm ${PREFIX}/bin/code ~/.config/tmoe-linux/across_architecture_container.txt ${PREFIX}/bin/startx11vnc 2>/dev/null || sudo rm -rfv "${DEBIAN_FOLDER}" ${PREFIX}/bin/debian ${PREFIX}/bin/startvnc ${PREFIX}/bin/stopvnc ${PREFIX}/bin/startxsdl ${PREFIX}/bin/debian-rm ${PREFIX}/bin/code ~/.config/tmoe-linux/across_architecture_container.txt ${PREFIX}/bin/startx11vnc 2>/dev/null
-	if [ -d "${HOME}/debian_armhf" ]; then
-		echo "检测到疑似存在树莓派armhf系统，正在移除..."
-		chmod 777 -R "${HOME}/debian_armhf"
-		rm -rf "${HOME}/debian_armhf" 2>/dev/null || sudo rm -rfv "${HOME}/debian_armhf"
-	fi
-	sed -i '/alias debian=/d' ${PREFIX}/etc/profile
-	sed -i '/alias debian-rm=/d' ${PREFIX}/etc/profile
-	source profile >/dev/null 2>&1
-	echo 'The container has been removed. If you want to uninstall aria2, enter "apt remove aria2" or "apt purge aria2"'
-	echo '移除完成，如需卸载aria2,请手动输apt remove aria2'
-	echo '其它相关依赖，如pv、dialog、procps、proot、wget等，均需手动卸载。'
-	echo 'If you want to reinstall, it is not recommended to remove the image file.'
+	read opt
+	case $opt in
+	y* | Y* | "")
+		chmod 777 -R ${DEBIAN_FOLDER}
+		rm -rfv "${DEBIAN_FOLDER}" ${PREFIX}/bin/debian ${PREFIX}/bin/startvnc ${PREFIX}/bin/stopvnc ${PREFIX}/bin/startxsdl ${PREFIX}/bin/debian-rm ${PREFIX}/bin/code ~/.config/tmoe-linux/across_architecture_container.txt ${PREFIX}/bin/startx11vnc 2>/dev/null || sudo rm -rfv "${DEBIAN_FOLDER}" ${PREFIX}/bin/debian ${PREFIX}/bin/startvnc ${PREFIX}/bin/stopvnc ${PREFIX}/bin/startxsdl ${PREFIX}/bin/debian-rm ${PREFIX}/bin/code ~/.config/tmoe-linux/across_architecture_container.txt ${PREFIX}/bin/startx11vnc 2>/dev/null
+		if [ -d "${HOME}/debian_armhf" ]; then
+			echo "检测到疑似存在树莓派armhf系统，正在移除..."
+			chmod 777 -R "${HOME}/debian_armhf"
+			rm -rf "${HOME}/debian_armhf" 2>/dev/null || sudo rm -rfv "${HOME}/debian_armhf"
+		fi
+		sed -i '/alias debian=/d' ${PREFIX}/etc/profile
+		sed -i '/alias debian-rm=/d' ${PREFIX}/etc/profile
+		source profile >/dev/null 2>&1
+		echo 'The debian system has been removed. If you want to uninstall aria2, enter "apt remove aria2" or "apt purge aria2"'
+		echo '移除完成，如需卸载aria2,请手动输apt remove aria2'
+		echo "Deleted已删除"
+		;;
+	n* | N*) echo "skipped." ;;
+	*) echo "Invalid choice. skipped." ;;
+	esac
 	echo "若需删除tmoe-linux管理器，则请输rm -f ${PREFIX}/bin/debian-i"
+	echo 'If you want to reinstall, it is not recommended to remove the image file.'
 	echo "${YELLOW}若您需要重装容器，则不建议删除镜像文件。${RESET} "
 	#ls -lh ~/debian-sid-rootfs.tar.xz 2>/dev/null
 	#ls -lh ~/debian-buster-rootfs.tar.xz 2>/dev/null
